requirements.txt
Cython
torch
torchvision
torchaudio
numpy>=1.22.0
scipy
regex
cn2an
opencc
inflect
unidecode
eng_to_ipa
pyopenjtalk-prebuilt
jamo
ko_pron
pypinyin
jieba
indic_transliteration
num_thai
pydub
numba
tencentcloud-sdk-python-tmt
langdetect
